Lazy-load entities in MetadataViewer and Visualizer
