+++ src/h5web/providers/Provider.tsx
@@ -1,9 +1,6 @@
 import { ReactElement, ReactNode, useMemo } from 'react';
-import { useAsync } from 'react-use';
 import { createFetchStore } from 'react-suspense-fetch';
 import { ProviderAPI, ProviderContext } from './context';
-import ErrorMessage from '../visualizer/ErrorMessage';
-import styles from '../visualizer/Visualizer.module.css';
 import type { Entity } from './models';
 import { isGroup } from '../guards';
 
@@ -14,11 +11,6 @@ interface Props {
 
 function Provider(props: Props): ReactElement {
   const { api, children } = props;
-
-  // Wait until metadata is fetched before rendering app
-  const { value: metadata, error } = useAsync(async () => {
-    return api.getMetadata();
-  }, [api]);
 
   const entitiesStore = useMemo(() => {
     const childCache = new Map<string, Entity>();
@@ -34,7 +26,7 @@ function Provider(props: Props): ReactElement {
         // Cache non-group children (datasets, datatypes and links)
         entity.children.forEach((child) => {
           if (!isGroup(child)) {
-            childCache.set(path, entity);
+            childCache.set(child.path, child);
           }
         });
       }
@@ -50,19 +42,10 @@ function Provider(props: Props): ReactElement {
     return createFetchStore(api.getValue.bind(api));
   }, [api]);
 
-  if (error) {
-    return <ErrorMessage error={error} />;
-  }
-
-  if (!metadata) {
-    return <p className={styles.fallback}>Loading...</p>;
-  }
-
   return (
     <ProviderContext.Provider
       value={{
         domain: api.domain,
-        metadata,
         entitiesStore,
         valuesStore,
       }}
