+++ src/h5web/visualizer/Visualizer.test.tsx
@@ -82,9 +82,18 @@ describe('Visualizer', () => {
     expect(tabs[0]).toHaveTextContent(Vis.NxImage);
   });
 
-  test('show "NX Image" tab for NXentry group with 2D default signal', async () => {
+  test('show "NX Image" tab for NXentry group with relative path to 2D default signal', async () => {
     renderApp();
     await selectExplorerNode('nexus_entry');
+
+    const tabs = await findVisSelectorTabs();
+    expect(tabs).toHaveLength(1);
+    expect(tabs[0]).toHaveTextContent(Vis.NxImage);
+  });
+
+  test('show "NX Image" tab for NXentry group with absolute path to 2D default signal', async () => {
+    renderApp();
+    await selectExplorerNode('nexus_entry/nx_process/absolute_default_path');
 
     const tabs = await findVisSelectorTabs();
     expect(tabs).toHaveLength(1);
@@ -145,7 +154,12 @@ describe('Visualizer', () => {
 
   test('show error when encountering malformed NeXus metadata', async () => {
     renderApp();
+
+    const { consoleMock, resetConsole } = mockConsoleMethod('error');
     await selectExplorerNode('nexus_malformed');
+
+    await selectExplorerNode('default_not_string');
+    expect(await screen.findByText(/to be a string/u)).toBeVisible();
 
     await selectExplorerNode('default_not_found');
     expect(await screen.findByText(/entity at path/u)).toBeVisible();
@@ -155,5 +169,8 @@ describe('Visualizer', () => {
 
     await selectExplorerNode('signal_not_found');
     expect(await screen.findByText(/to exist/u)).toBeVisible();
+
+    expect(consoleMock).toHaveBeenCalledTimes(8); // React logs two stack traces per error
+    resetConsole();
   });
 });
